+++ lib/crawler/fetcher/recorder.ex
@@ -4,16 +4,16 @@ defmodule Crawler.Fetcher.Recorder do
   @doc """
   ## Examples
 
-      iex> Recorder.record(url: "url1", level: 2)
-      {:ok, [level: 3, url: "url1"]}
+      iex> Recorder.record(url: "url1", depth: 2)
+      {:ok, [depth: 3, url: "url1"]}
 
-      iex> Recorder.record(url: "url2", level: 2)
+      iex> Recorder.record(url: "url2", depth: 2)
       iex> Store.find("url2")
       %Page{url: "url2"}
   """
   def record(opts) do
     with {:ok, _pid} <- store_url(opts),
-         opts        <- store_url_level(opts)
+         opts        <- store_url_depth(opts)
     do
       {:ok, opts}
     end
@@ -27,7 +27,7 @@ defmodule Crawler.Fetcher.Recorder do
     Store.add(opts[:url])
   end
 
-  defp store_url_level(opts) do
-    Keyword.replace(opts, :level, opts[:level] + 1)
+  defp store_url_depth(opts) do
+    Keyword.replace(opts, :depth, opts[:depth] + 1)
   end
 end